#201 Extract functionality of onInViewChange to a component
