+++ src/components/article/ArticleTitle.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'gatsby';
 
-import HeaderOne from './HeaderOne';
+import HeaderOne from './headers/HeaderOne';
 
 import './articleTitle.module.scss';
